Setup documenter
