+++ config/api-documenter.json
@@ -1,0 +1,5 @@
+{
+  "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-documenter.schema.json",
+  "outputTarget": "markdown",
+  "newlineKind": "lf"
+}
