+++ scripts/watch.js
@@ -1,4 +1,11 @@
+const ts = require("typescript");
+
 const { flags } = require("./common/flags");
-const { watcher } = require("./common/watcher");
+const { host } = require("./common/host");
+
+const watcher = ts.createSolutionBuilderWithWatch(host, ["tsconfig.json"], {
+  force: flags.force,
+  verbose: flags.verbose,
+});
 
 watcher.build(flags.project);
